minor refactor addressed pull request comment
