--- src/ui-spa/src/common/utils/getSuspectsWithNoCharges.ts
@@ -1,7 +1,0 @@
-import { type SuspectFormData } from "../reducers/caseRegistrationReducer";
-
-export const getSuspectsWithNoCharges = (
-  suspects: SuspectFormData[],
-): SuspectFormData[] => {
-  return suspects.filter((suspect) => suspect.charges.length === 0);
-};
