+++ src/ui-spa/src/components/case-registration/case-monitoring-codes/index.tsx
@@ -13,7 +13,7 @@ import { getCaseMonitoringCodes } from "../../../apis/gateway-api";
 import { useQuery } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
 import { type CaseMonitoringCodes } from "../../../schemas";
-import { getSuspectsWithNoCharges } from "../../../common/utils/getSuspectsWithNoCharges";
+import { isMonitoringCodeOptional } from "../../../common/utils/isMonitoringCodeOptional";
 import useChargesCount from "../../../common/hooks/useChargesCount";
 import pageStyles from "./index.module.scss";
 import styles from "../index.module.scss";
@@ -52,11 +52,8 @@ const CaseMonitoringCodesPage = () => {
   });
 
   const isOptional = useMemo(() => {
-    const noChargeSuspects = getSuspectsWithNoCharges(state.formData.suspects);
-    return (
-      !noChargeSuspects.length && state.formData.suspectDetailsRadio === "yes"
-    );
-  }, [state.formData.suspectDetailsRadio, state.formData.suspects]);
+    return isMonitoringCodeOptional(state.formData.suspects);
+  }, [state.formData.suspects]);
 
   const [formDataErrors, setFormDataErrors] = useState<FormDataErrors>({});
